requirements.txt
python-telegram-bot[job-queue]==21.0
python-whois==0.9.4
aiohttp==3.9.1
